[03275] Split URL validation tests from utils.test.ts into url.test.ts
Move validateRedirectUrl, validateLinkUrl, media validation (validateImageUrl,
validateAudioUrl, validateVideoUrl), and validateEmbedUrl test suites from
utils.test.ts into url.test.ts alongside existing getFullUrl tests.

utils.test.ts now only contains getIvyHost and getShellParam tests.
